Reorganize tests
The directory test/ did not use correctly dune dependencies. It is
better to limit the dependencies of each test, so dune won't rerun it
if you don't modify a sublibrary in lib/. Currently, the geneweb library
is very monolithic but we can improve this situation easily in
the future.

I also removed the target `make ci`. It is not a good practice to
disable tests on CI. A better practice consists in inverting the failed
test. So we can notice if a fix works.