Do not catch critical exceptions
Catching all the exception to print them is a bad pratice. You shoud
never catch Assert_failure for instance. If such an exception has been
raised, it means that an internal invariant of the program is broken and
the only appropriate behavior is to stop it completely.

If you catch it and print the exception, you will loss the backtrace! It
means you don't know where the failure occurs in your code base.

I only fix the problem in gwd.ml but I notice several positions in
wserver.ml where all the exception are reraised without the backtrace.