Use default flags of dune in dev profile
We should never disable warnings in the dev profile. Let Dune decide for
us what is the more appropriate.

The `DUNE_PROFILE` variable environment is known by `dune build` so we
do not need to add `--profile $(DUNE_PROFILE)` everywhere in the
  Makefile.

The CI should compile the project in dev profile and release profile.
Currently, we only compile it in release profile, which means that
warnings are not fatal. After this commit, we compile it with both profiles.