Force Dune to build the Geneweb library before gwrepl
The previous dune file assumes that the presence of `.depend` was
sufficient to ensure that the rule to generate `data.cppo.ml` won't run
before the cma of Geneweb library are available in `_build`. It was
wrong! The CI failed sometimes because `.depend` is produced to early.

This commit adds an extra hack on this hackish dune file. I add an
explicit dependency on the rule for `data.cppo.ml` which ensures that the cma
of geneweb have been built.