Add a gzip input_line buffered
Unfortunately, the camlzip library does not expose a convenient function
as `In_channel.input_line`. We do not want to read all the file in a
single string before splitting it on newline character. The file of the
uncompressed file can be very large.

This commit adds an small wrapper around `Gzip.in_channel` with
the capability of the read line per line the file. In the worst case,
the new function `input_line` use O(n) memory space where n is the size
of the current line.