Add buffered `input_line` for Camlzip in_channel
The input_channel type exposed by Camlzip does not provide a convenient
API as stdlib's in_channel. In particular, there is no `input_line`
function to read a gzip file line by line.

For small gzip file, a simple workaround is to read the entire
file in one pass and store it in an OCaml string. However, this approach
as two drawbacks:
1. On 32-bits architecture, the size of OCaml string is really limited
   (around 16MB), which means we cannot read large gzip file.
2. The entire file must be stored in memory, even if only a single
   line in memory is needed to perform a task.

This PR addresses this issue by implementing a small wrapper around the
Camlzip in_channel type. The `My_gzip.input_line` function limits memory usage
to the length of the next line.